feat: 장소 입력 구조 수정
